multi_modal_generator/scripts/requirements.txt
flask
torch
transformers
Pillow
numpy
opencv-python
openai
librosa
soundfile
